Move Snapshot from legacy to new core module.
