+++ kotlin/workflow-host/src/main/java/com/squareup/workflow/internal/SubtreeManager.kt
@@ -15,13 +15,13 @@
  */
 package com.squareup.workflow.internal
 
-import com.squareup.workflow.legacy.Snapshot
-import com.squareup.workflow.legacy.parse
-import com.squareup.workflow.legacy.readByteStringWithLength
+import com.squareup.workflow.Snapshot
+import com.squareup.workflow.parse
+import com.squareup.workflow.readByteStringWithLength
 import com.squareup.workflow.Workflow
 import com.squareup.workflow.WorkflowAction
 import com.squareup.workflow.internal.Behavior.WorkflowOutputCase
-import com.squareup.workflow.legacy.writeByteStringWithLength
+import com.squareup.workflow.writeByteStringWithLength
 import kotlinx.coroutines.experimental.selects.SelectBuilder
 import kotlin.coroutines.experimental.CoroutineContext
